# [Draw][Breakpoint] Fix debug breakpoints to work in placement and routing