--- a/vpr/src/route/route.cpp
+++ b/vpr/src/route/route.cpp
@@ -12,6 +12,7 @@
 #include "rr_graph.h"
 #include "router_lookahead_report.h"
 #include "vtr_time.h"
+#include "vtr_expr_eval.h"
 
 bool route(const Netlist<>& net_list,
            int width_fac,
@@ -251,6 +252,10 @@ bool route(const Netlist<>& net_list,
     int rcv_finished_count = RCV_FINISH_EARLY_COUNTDOWN;
 
     print_route_status_header();
+#ifndef NO_GRAPHICS
+    // Reset router iteration in the current route attempt.
+    get_bp_state_globals()->get_glob_breakpoint_state()->router_iter = 0;
+#endif
     for (itry = 1; itry <= router_opts.max_router_iterations; ++itry) {
         /* Reset "is_routed" and "is_fixed" flags to indicate nets not pre-routed (yet) */
         for (auto net_id : net_list.nets()) {
@@ -267,6 +272,11 @@ bool route(const Netlist<>& net_list,
         if (budgeting_inf.if_set()) {
             worst_negative_slack = timing_info->hold_total_negative_slack();
         }
+
+#ifndef NO_GRAPHICS
+        // Update router information and check breakpoint.
+        update_router_info_and_check_bp(BP_ROUTE_ITER, -1);
+#endif
 
         /* Initial criticalities: set to 1 on the first iter if the user asked for it */
         if (router_opts.initial_timing == e_router_initial_timing::ALL_CRITICAL && itry == 1)
@@ -400,9 +410,6 @@ bool route(const Netlist<>& net_list,
         if (legal_convergence_count >= router_opts.max_convergence_count
             || iter_results.stats.connections_routed == 0
             || early_reconvergence_exit_heuristic(router_opts, itry_since_last_convergence, timing_info, best_routing_metrics)) {
-#ifndef NO_GRAPHICS
-            update_router_info_and_check_bp(BP_ROUTE_ITER, -1);
-#endif
             break; //Done routing
         }
 
@@ -410,9 +417,6 @@ bool route(const Netlist<>& net_list,
          * Abort checks: Should we give-up because this routing problem is unlikely to converge to a legal routing?
          */
         if (itry == 1 && early_exit_heuristic(router_opts, wirelength_info)) {
-#ifndef NO_GRAPHICS
-            update_router_info_and_check_bp(BP_ROUTE_ITER, -1);
-#endif
             //Abort
             break;
         }
@@ -423,18 +427,12 @@ bool route(const Netlist<>& net_list,
 
             if (!std::isnan(est_success_iteration) && est_success_iteration > abort_iteration_threshold && router_opts.routing_budgets_algorithm != YOYO) {
                 VTR_LOG("Routing aborted, the predicted iteration for a successful route (%.1f) is too high.\n", est_success_iteration);
-#ifndef NO_GRAPHICS
-                update_router_info_and_check_bp(BP_ROUTE_ITER, -1);
-#endif
                 break; //Abort
             }
         }
 
         if (itry == 1 && router_opts.exit_after_first_routing_iteration) {
             VTR_LOG("Exiting after first routing iteration as requested\n");
-#ifndef NO_GRAPHICS
-            update_router_info_and_check_bp(BP_ROUTE_ITER, -1);
-#endif
             break;
         }