--- a/vpr/src/route/route_utils.cpp
+++ b/vpr/src/route/route_utils.cpp
@@ -673,16 +673,14 @@ void update_draw_pres_fac(const float /*new_pres_fac*/) {
 #ifndef NO_GRAPHICS
 void update_router_info_and_check_bp(bp_router_type type, int net_id) {
     t_draw_state* draw_state = get_draw_state_vars();
-    if (!draw_state->list_of_breakpoints.empty()) {
-        if (type == BP_ROUTE_ITER)
-            get_bp_state_globals()->get_glob_breakpoint_state()->router_iter++;
-        else if (type == BP_NET_ID)
-            get_bp_state_globals()->get_glob_breakpoint_state()->route_net_id = net_id;
-        f_router_debug = check_for_breakpoints(false);
-        if (f_router_debug) {
-            breakpoint_info_window(get_bp_state_globals()->get_glob_breakpoint_state()->bp_description, *get_bp_state_globals()->get_glob_breakpoint_state(), false);
-            update_screen(ScreenUpdatePriority::MAJOR, "Breakpoint Encountered", ROUTING, nullptr);
-        }
+    if (type == BP_ROUTE_ITER)
+        get_bp_state_globals()->get_glob_breakpoint_state()->router_iter++;
+    else if (type == BP_NET_ID)
+        get_bp_state_globals()->get_glob_breakpoint_state()->route_net_id = net_id;
+    f_router_debug = check_for_breakpoints(false);
+    if (f_router_debug) {
+        breakpoint_info_window(get_bp_state_globals()->get_glob_breakpoint_state()->bp_description, *get_bp_state_globals()->get_glob_breakpoint_state(), false);
+        update_screen(ScreenUpdatePriority::MAJOR, "Breakpoint Encountered", ROUTING, nullptr);
     }
 }
 #endif