This PR fixes the debug breakpoint so that breakpoints are triggered under the intended conditions in both placement and routing.

- Previously, the router iteration breakpoint was not triggered correctly because the counter was not being updated properly.
- In addition, if placement breakpoints were still active during routing, none of the routing breakpoints would trigger since they weren’t being guarded against.


Both issues are addressed in this PR.

#### Related Issue
See comments in the PR #3278