@haydar-c 2025-09-26
Based on comments, I ran the largest 2 VTR benchmarks on my desktop and putting CPU time comparisons here.

### Build: **No Graphics**

| Circuit       | Normalized Placement Time (With Change / Without Change) | Normalized Placement Time (With Change / Without Change) |
|---------------|-----------------------------------------------------------|--------------|
| mcml.v              |  1.001724667  | 1.005574136  |
|LU32PEEng.v    |  1.009246933 | 1.001697073  |


### Build: **With Graphics** (`--disp off`)
| Circuit       | Normalized Placement Time (With Change / Without Change) | Normalized Placement Time (With Change / Without Change) |
|---------------|-----------------------------------------------------------|--------------|
| mcml.v              |  1.008656351  | 1.015185602  |
|LU32PEEng.v    |  1.016414734   | 1.005959983  |

The latest change shows a small increase in runtime, but it’s marginal. I’ll defer to the reviewer on whether we should optimize it for CPU time.