@haydar-c 2025-09-26
Yes, the second columns is meant to be normalized routing time. Sorry for the confusion, updated now.